fix validation message showing different 'value'
Validation functions use 'ctxt.point' to validate the input, yet the inserts sent to the template will render either ctxt.value (if it's truthy) or killcircled, JSON-stringified ctxt.point. This is problematic because (1) ctxt.value may be 0, which is falsy and ignored and (2) the stringified point can vary from the actually-compared point. For instance, validation may run against a point of `-Infinity` but `null` would passed in the validation message because `JSON.stringify(-Infinity)` is 'null'.